feat(Views): implement first iteration of dependency trackers
A dependency tracker is used with connect. It is a wrapper for computed, meaning that extraction of state and props inside the computed is tracked by the component and merged
into its own dependency map. The components automatically checks for need to recalculate a computed by comparing flush change map or props change with each computed dependency map

BREAKING CHANGE:
- Use new compute API instead of Computed
- Functions not longer available as first argument of connect, use tags and/or compute
- Use tags at first argument of connect to bind signals
- mergeProps callback signature now is `(depsProps, ownProps, resolve) => newProps`
- bigComponentWarning is now just integer and equals 10 by default